python_engine/requirements.txt
google-generativeai
newspaper4k
gspread
google-auth
requests
pandas
lxml_html_clean
python-dotenv
